requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.9.1
asyncpg==0.30.0
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.1.1
dotenv==0.9.9
flake8==7.1.1
flake8-docstrings==1.7.0
frozenlist==1.7.0
google-auth==2.41.1
google-auth-oauthlib==1.2.2
greenlet==3.2.4
gspread==6.2.1
idna==3.10
iniconfig==2.1.0
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.6.4
oauthlib==3.3.1
packaging==25.0
pep8-naming==0.14.1
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pydantic==2.11.7
pydantic_core==2.33.2
pydocstyle==6.3.0
pyflakes==3.2.0
pyTelegramBotAPI==4.22.1
pytest==7.1.3
pytest-django==4.9.0
pytest-lazy-fixture==0.6.3
pytest-subtests==0.13.1
python-dotenv==1.1.1
pytils==0.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
snowballstemmer==3.0.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
